requirements.txt
discord.py
asyncpg
python-dotenv
langchain
langchain-community
langchain-openai
openai
tiktoken